correct strlcpy defns

+++ src/sys/lib/libsa/stand.h
@@ -1,4 +1,4 @@
-/**	$MirOS: src/sys/lib/libsa/stand.h,v 1.4 2006/06/16 21:52:54 tg Exp $	*/
+/**	$MirOS: src/sys/lib/libsa/stand.h,v 1.5 2006/11/21 02:53:39 tg Exp $	*/
 /*	$OpenBSD: stand.h,v 1.43 2004/01/03 14:08:53 espie Exp $	*/
 /*	$NetBSD: stand.h,v 1.18 1996/11/30 04:35:51 gwr Exp $	*/
 
@@ -151,10 +151,10 @@ int	memcmp(const void *, const void *, size_t);
 char	*strncpy(char *, const char *, size_t);
 int	strncmp(const char *, const char *, size_t);
 int	strcmp(const char *, const char *);
-size_t strlcpy(char *, const char *, int)
-	__attribute__ ((__bounded__(__string__,1,3)));
-size_t strlcat(char *, const char *, int)
-	__attribute__ ((__bounded__(__string__,1,3)));
+size_t	 strlcpy(char *, const char *, size_t)
+		__attribute__ ((__bounded__(__string__,1,3)));
+size_t	 strlcat(char *, const char *, size_t)
+		__attribute__ ((__bounded__(__string__,1,3)));
 size_t	strlen(const char *);
 long	strtol(const char *, char **, int);
 char	*strchr(const char *, int);
